add whatwg-fetch to fix fetch errors in tests and resolve addtocart unmounted component warning by updating useEffect
