+++ package.json
@@ -32,7 +32,8 @@
     "react-redux": "^7.2.5",
     "react-router-dom": "^5.3.0",
     "regenerator-runtime": "^0.13.9",
-    "style-loader": "^3.2.1"
+    "style-loader": "^3.2.1",
+    "whatwg-fetch": "^3.6.2"
   },
   "devDependencies": {
     "@babel/core": "^7.15.5",
