+++ tests/ProductInfo.test.jsx
@@ -4,6 +4,7 @@
 
 // Product Info Tests ==============================================
 import React from 'react';
+import 'whatwg-fetch';
 import {
   render, cleanup, waitFor, fireEvent, screen,
 } from '@testing-library/react';
@@ -13,8 +14,12 @@ import axios from 'axios';
 import { Provider } from 'react-redux';
 import ProductInfo from '../client/src/components/Overview/ProductInfo';
 import Style from '../client/src/components/Overview/Style';
-
 import store from '../client/src/store/store';
+import mockData from './fixtures/OverviewMockData';
+
+const {
+  mockProductData, mockStyleData, mockStyle, mockCartData,
+} = mockData;
 
 beforeAll(() => {
   axios.get.mockImplementation((url) => {
@@ -35,36 +40,6 @@ jest.mock('axios');
 jest.mock('../client/src/components/Overview/overview.css', () => () => (<div>Placeholder Overview Style</div>));
 jest.mock('../client/src/components/Overview/Price.jsx', () => () => (<div>$100</div>));
 jest.mock('../client/src/components/Overview/StyleSelector.jsx', () => () => (<div>Styles Placeholder</div>));
-
-const mockProductData = {
-  id: 3,
-  name: 'Morning Joggers',
-  slogan: 'Make yourself a morning person',
-  description: "Whether you're a morning person or not. Whether you're gym bound or not. Everyone looks good in joggers.",
-  category: 'Pants',
-  default_price: '40',
-};
-
-const mockStyleData = {
-  product_id: '1',
-  results: [
-    {
-      style_id: 293480,
-      name: 'Desert Brown & Tan',
-      original_price: '140.00',
-    },
-    {
-      style_id: 123456,
-      name: 'Testing only',
-      original_price: '0',
-    },
-    {
-      style_id: 654321,
-      name: 'Testing 2',
-      original_price: '10.00',
-    },
-  ],
-};
 
 // TESTS =======================================================
 it('should load and display the selected product data',
@@ -110,22 +85,3 @@ it('should load and display the selected product data', async () => {
     .catch((err) => console.log(err));
 });
 
-// Style Selector Tests ==============================================
-test('selected images should have select formatting with border and checkmark', () => {
-  const style = {
-    style_id: 123456,
-    name: 'Selected Style',
-    original_price: '0',
-    photos: [
-      { thumbnail_url: 'https://testing.com' },
-    ],
-  };
-  const { getByAltText } = render(
-    <Provider store={store}>
-      <Style style={style} />
-    </Provider>,
-  );
-
-  fireEvent.click(getByAltText('Selected Style'));
-  expect(getByAltText('Selected Style')).toHaveClass('overview-style-selected');
-});
